Refactor district report into separate component
